requirements/base.txt
anyio==4.7.0
asgiref==3.8.1
async-timeout==5.0.1
betterproto==2.0.0b1
certifi==2024.12.14
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.4
django-debug-toolbar==4.4.6
ffmpy==0.5.0
grpclib==0.3.2
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
icalendar==6.1.0
idna==3.10
mashumaro==3.15
multidict==6.1.0
packaging==24.2
pillow==11.0.0
protobuf==5.29.2
protobuf3-to-dict==0.1.5
psycopg2-binary==2.9.10
pyee==12.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-socks==2.5.3
pytz==2024.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
TikTokLive==6.3.0.post2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
websockets==14.1
websockets_proxy==0.1.3
